Moved StoredMixnodeBond to storage.rs
